Deploy only if secure environment variables are available.
Added a conditional test in the `after_success` script to test if there are
secure environment variables. Adding a `if` bash statement will make the script
to exit with 0 even if some internal command fails. Adding `travis_terminate 1`
terminates the build if an internal command fails.

Inspired by http://steven.casagrande.io/articles/travis-ci-and-if-statements/